requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
python-dotenv==1.0.1
requests==2.28.1
schedule==1.1.0
tomli==2.0.1
urllib3==1.26.13
